move tests to separate folder
