+++ alien_invasion/scoreboard.py
@@ -1,6 +1,6 @@
 import pygame.font
 from pygame.sprite import Group
-from ship import Ship
+from alien_invasion.ship import Ship
 
 class Scoreboard:
     """A class to report scoring information."""
