+++ tests/test_bullet.py
@@ -1,6 +1,6 @@
-from alien_invasion import AlienInvasion
-from alien import Alien
-from bullet import Bullet, YellowBullet, RedBullet
+from alien_invasion.alien_invasion import AlienInvasion
+from alien_invasion.alien import Alien
+from alien_invasion.bullet import Bullet, YellowBullet, RedBullet
 
 class TestBullet:
