+++ tests/test_explosion.py
@@ -1,6 +1,6 @@
-from alien_invasion import AlienInvasion
-from alien import Alien
-from explosion import Explosion
+from alien_invasion.alien_invasion import AlienInvasion
+from alien_invasion.alien import Alien
+from alien_invasion.explosion import Explosion
 import unittest.mock as mock
 import pygame
