+++ tests/test_settings.py
@@ -1,4 +1,4 @@
-from alien_invasion import AlienInvasion
+from alien_invasion.alien_invasion import AlienInvasion
 
 
 class TestSettings:
